requirements.txt
anyio==3.6.1
autopep8==1.7.0
certifi==2022.6.15
et-xmlfile==1.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
numpy==1.23.2
openpyxl==3.0.10
pandas==1.4.3
pyactiveresource==2.2.2
pycodestyle==2.9.1
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
rfc3986==1.5.0
ShopifyAPI==12.0.0
six==1.16.0
sniffio==1.2.0
toml==0.10.2
